api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["HabitUp <agilefans@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
uvicorn = "^0.31.0"
fastapi = "^0.115.0"
sqlmodel = "^0.0.22"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
